Path_following_Code
This is the code that allows the robot to sense a line and follow it. This is done by using the Infrared Sensor. The way the sensor works is by releasing an infrared signal which is then reflected of of a surface back to the sensor. The track that the robot is set to follow is made of black electrical tape which does not reflect a signal back. The robot then senses this and changes the speed of the motor based on which of the five sensors is deactivated.